pyproject.toml
[project]
name = "zippee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bcrypt>=4.3.0",
    "flasgger>=0.9.7.1",
    "flask>=3.1.2",
    "flask-pydantic>=0.13.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
]
